avoid bogus virtio devices in device listing (bsc #960507)
The code actually creates two entries for each device: one is a pci device
and the other a virtio device. We want only the second because only this has
the interesting attributes like interface, mac address, or can be directly
linked to underlying block devices.

The patch keeps both entries internally but tags the pci device as having an
'unknown' device class effectively preventing it from showing up in any
device listing.