add support for mmc/sdio devices (bsc #996183)
This adds devices found on the mmc and sdio bus in sysfs.

For this it's also necessary to introduce a new id type 'sdio' in libhd as
sdio device ids are separate from, say, pci or usb ids.

The patch additionally recognizes the bus controllers and network devices;
storage devices were already handled.

If you are interested in technical details, specs are available at
https://www.sdcard.org/developers/overview/sdio.