Don't assume that "/var/lib/hardware" exists on all platforms
Ubuntu, for instance does not have the hardware directory
/var/lib/hardware where configs are saved by default and it
fails to save the config files so we need to ensure that "dir"
exists as well along with the extended dir(with udi attached).

I have implemented a "mkdir -p" as mkdir_p() that takes a bool
to tell if we want to have the parent directories created as
well apart from the path and mode arguments.

This patch allows the user to save the config when run as root.
If non-root users also need the capability that will be another
patch.

Signed-off-by: Chandni Verma <chandni@linux.vnet.ibm.com>