Keep NVMe's namespace output consistency when nvme_core.multipath=1 (bsc#1199948)
When nvme_core.multipath=1 the device link in the SysFS's namespace entry
points to nvme-subsys# and that causes some missing information,
specificaly about PCI and namespace's parent device.

Here we go through the parent device withihn the nvme-subsys to gather
complete information, and making the output's compatible to the previously
seen.

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>